+++ src/util.hpp
@@ -17,7 +17,7 @@ namespace util
     template <typename T>
     bool isSorted(const std::vector<T>& v)
     {
-        for (int i = 0; i < v.size() - 1; ++i)
+        for (size_t i = 0; i < v.size() - 1; ++i)
         {
             if (v[i] > v[i + 1])
             {
